Generate weekLists routes dynamically from base routes
Rationale: for each new resource (for example for recipes) a bunch of basic resource endpoints had to be written for basic functionality such as listing all, creating new, updating and deleting records. This resulted in a lot of duplicated code (and all problems that come with that).

Now these endpoints can be dynamically generated from base functions with relatively little code and while still allowing for creating fully custom endpoints for each route as before.

The main idea is just to use partial functions and take advantage of the fact that decorators are just functions that take other functions as arguments.

Flask does have some support for this use case but that seemed to still require more boilerplate and the whole idea is to keep boilerplate to a minimum.