Split bottomSection style and add proper sort button
Rationale: style split so button can be part of the bottomSection but styled separately.

And button and split some weekList functions as we want to call only some of them in the case we use the sumbmit button (filter + optionally sort + submit), the sort button in view mode (filter + sort + submit) and in edit mode (filter + sort).