Remove shops GET, PUT, DELETE routes and add location routes
Rationale: for shops we currently only use get all and the PUT & DELETE routes for need some extra work over the base versions so would currently actually leave the db in a broken state.

For the locations I added routes to be able to add and delete locations properly, currently only to be used directly by a dev/admin.