Fix weekList delete button overflowing on safari
Seems like the problem might be the font-family. This was manually set to `Segoe UI` (I guess because I copied it from the nav logout button which I might have styled this way previously to match the links in the navbar).  Segoe UI is a microsoft font so probably isn't shipped on MacOS. Removing that declaration fixes the issue and doesn't seem to affect Firefox or Chrome in any way (the logout button neither).